requirements.txt
psycopg2-binary==2.9.1
sqlalchemy
streamlit==1.0.0
pandas
numpy
matplotlib
altair
markdown
modules
webbrowser
awesome_streamlit==0.5.1
markdown
